+++ md-roam.el
@@ -5,8 +5,8 @@
 ;; Author: Noboru Ota <https://github.com/nobiot>, <https://gitlab.com/nobiot>
 ;; Maintainer: Noboru Ota <me@nobiot.com>
 ;; Created: April 15, 2020
-;; Modified: June 19, 2020
-;; Version: 1.3.1
+;; Modified: July 5, 2020
+;; Version: 1.4.0
 ;; Keywords:
 ;; Homepage: https://github.com/nobiot/md-roam, https://gitlab.com/nobiot/md-roam
 ;; Package-Requires: ((emacs 26.3) (dash) (s) (f) (org-roam))
@@ -61,6 +61,10 @@
   ;; Assumed to be case insensitive
   "\\(^.*ROAM_ALIAS:[ \t]*\\)\\(.*\\)")
 
+(defvar md-roam-regex-ref-key
+  ;; Assumed to be case insensitive
+  "\\(^.*ROAM_KEY:[ \t]*\\)\\(.*\\)")
+
 (defvar md-roam-regex-headline
   (concat "^\s*\n"                   ;exludes YAML front matter
           "\\(.*$\\)\n\\(^[=-]+$\\)" ;heading with '=' and '-'
@@ -95,18 +99,70 @@
   "\\(?:[^[:alnum:]]\\|^\\)\\(-?@\\)\\([-a-zA-Z0-9_+:]+\\)"
   "Regular expression for stand-alone citation with no anchor.")
 
+;;; Regexp for extracting tags
+;;  This regexp is intended to be compatible with Zettlr:
+;;
+;;     #tag1 #tag-with-hyphen #tag_with_underscore
+;;
+;;  Note that iA Writer treats hyphen "-" as a word delimiter
+;;  within a tag. That is, iA Writer treats #tag-hyphen tagged as
+;;  #tag, and ignores "-hyphen".
+;;
+;;  If iA Writer's stricter style is preferred, the regexp should
+;;  be defined as:
+;;
+;;     "\\([^/s]\\)\\([#@][[:alnum:]_]+\\)"
+;;
+
+(defvar md-roam-regex-tags-zettlr-style
+  "\\([^/s]\\)\\([#@][[:alnum:]_-]+\\)")
+
 (defvar md-roam-verbose t)
 
 ;;; Md-roam customizing
 
 (defcustom md-roam-file-extension-single "md"
-  "Defines the extesion to be used for md-roam wihtin org-roam.
+  "Defines the extesion to be used for Md-roam within Org-roam directory.
 Unlike 'org-roam-file-extension', this is a single value, not a list.
-It is assumed to be a markdown file extension, e.g. .md, and .markdown."
-  :type '(repeat string)
+It is intended to be used for you to define a different markdown extension,
+such as .md and .markdown."
+
+  :type 'string
+  :group 'org-roam)
+
+(defcustom md-roam-use-org-file-links t
+  "Defines if Md-roam extracts Org's file link for backlinks in md files.
+Default is t, that is to extract both wiki-links and Org's file links.
+For faster performance, set it to nil to extract only
+[[wiki-links]] of Md-roam and ignore the Org file links.
+This does not affect Org files within Org-roam directory."
+  
+  :type 'boolean
+  :group 'org-roam)
+
+(defcustom md-roam-use-org-extract-ref t
+  "Defines if Md-roam extracts REF_KEY using Org-roam logic.
+Default is t, that is to use #+ref_key with Org-roam logic.
+If nil, Md-roam uses its own regex to look for #+roam_key: or
+roam_key: within YAML front matter only.
+It's intended for better performance, and aesthetic style.
+Recommended if your bibliographic notes are written in markdown files.
+Leave it as default, if they are written in org files."
+
+  :type 'boolean
   :group 'org-roam)
 
 ;;; Md-roam functions
+
+
+;;;  Tell if a file is an .org file (or encrypted org file)
+(defun md-roam--org-file-p (path)
+  "Check if PATH is pointing to an org file.
+Return t or nil."
+  (let ((ext (org-roam--file-name-extension path)))
+    (when (string= ext "gpg")           ; Handle encrypted files
+      (setq ext (org-roam--file-name-extension (file-name-sans-extension path))))
+    (string= ext "org")))
 
 ;;;  Extracting title from markdown files (YAML frontmatter)
 ;;;  Add advice to org-roam--extract-and-format-titles
@@ -125,6 +181,21 @@ Return nil if the front matter does not exist, or incorrectly delineated by
          (endpoint (re-search-forward
                     md-roam-regex-yaml-font-matter-ending nil t 1)))
       (buffer-substring-no-properties startpoint endpoint))))
+
+(defun md-roam-get-yaml-front-matter-endpoint ()
+  "Return the endpoint of the YAML front matter of the current buffer.
+Return nil if the front matter does not exist, or incorrectly delineated by
+'---'. The front matter is required to be at the beginning of the file."
+
+  (save-excursion
+    (goto-char (point-min))
+    (when-let
+        ((startpoint (re-search-forward
+                      md-roam-regex-yaml-font-matter-beginning 4 t 1))
+         ;The beginning needs to be in the beginning of buffer
+         (endpoint (re-search-forward
+                    md-roam-regex-yaml-font-matter-ending nil t 1)))
+      endpoint)))
 
 (defun org-roam--extract-titles-mdtitle ()
   "Extract title from the current buffer (markdown file with YAML frontmatter).
@@ -258,13 +329,15 @@ FILE-PATH is mandatory as `org-roam--extract-links' identifies it."
 (defun md-roam--extract-links (original-extract-links &optional file-path)
   "Add markdown links (wiki and cite) for FILE-PATH to the org-roam equivalent.
 ORIGINAL-EXTRACT-LINKS is supplemented with md-roam functions.
-It should be used with 'advice-add'."
+It should be used with 'advice-add' and :around ."
   (let* ((file-path (or file-path
                         (file-truename (buffer-file-name))))
-         (links (apply original-extract-links file-path nil))
+         (links '())
          (md-links (md-roam--extract-wiki-links file-path))
          (md-cite-links (md-roam--extract-cite-links file-path)))
-
+    (when (or (md-roam--org-file-p file-path)
+              md-roam-use-org-file-links) ;For [[file:file.ext][desc]] within md
+      (setq links (apply original-extract-links file-path nil)))
     (when md-links
       (setq links (append md-links links)))
     (when md-cite-links
@@ -273,16 +346,39 @@ It should be used with 'advice-add'."
 
 (advice-add 'org-roam--extract-links :around #'md-roam--extract-links)
 
+;;; Md-roam extract ref via regex
+(defun md-roam--extract-ref-regex ()
+  "Extract roam_key from current buffer; return the type and the key.
+Use regex instead of `org-roam--extract-global-props'.
+Return cons of (type . key). Type is always 'file' for now."
+
+    (let ((frontmatter (md-roam-get-yaml-front-matter)))
+    (cond (frontmatter
+           (when (string-match md-roam-regex-ref-key frontmatter)
+             (cons "file" (match-string-no-properties 2 frontmatter)))))))
+
+(defun md-roam--extract-ref (original-extract-ref)
+  "Extract roam_key from current buffer.
+If current buffer is Org file, use ORIGINAL-EXTRACT-REF:
+`org-roam--extract-ref'.
+If not, use Md-roam specific regex to search within YAML front matter.
+It is meant to be used with `advice-add' :around."
+
+  (if md-roam-use-org-extract-ref
+      (funcall original-extract-ref)
+    (md-roam--extract-ref-regex)))
+
+(advice-add 'org-roam--extract-ref :around #'md-roam--extract-ref)
 
 ;;;; Adapt behaviour of org-roam-insert
 ;;;; Add advice to 'org-roam--format-link
 
 (defun md-roam--format-link (target &optional description)
   "Formats a [[wikilink]] for a given file TARGET and link DESCRIPTION.
-Add advice to 'org-roam--format-link' within 'org-roam-inert'.
-Customize `org-roam--file-name-extension' to define the extesion (e.g. md) that
-follow this behaviour."
-
+Add advice to 'org-roam--format-link' within 'org-roam-insert'.
+Customize `md-roam-file-extension-single' to define the extesion (e.g. md) that
+follows this behaviour."
+  
   (let ((ext (org-roam--file-name-extension (buffer-file-name (buffer-base-buffer)))))
     (if (string= ext md-roam-file-extension-single)
         (let* ((here (ignore-errors
@@ -300,6 +396,46 @@ follow this behaviour."
 
 (advice-add 'org-roam--format-link :before-until #'md-roam--format-link)
 
+(defun org-roam--extract-tags-md-buffer (_file)
+  "Extracts tags defined in the Zettlr style."
+  (save-excursion
+    (let (tags-list)
+      (goto-char (point-min))
+      (while (re-search-forward md-roam-regex-tags-zettlr-style nil t)
+        (let ((tag (match-string-no-properties 2)))
+          (when tag
+            (setq tags-list (append tags-list (list tag))))))
+      tags-list)))
+
+(defun org-roam--extract-tags-md-frontmatter (_file)
+  "Extracts tags defined in the Zettlr style only within front matter."
+
+  (let ((endpoint (md-roam-get-yaml-front-matter-endpoint)))
+    (cond (endpoint
+           (save-excursion
+             (let (tags-list)
+               (goto-char (point-min))
+               (while (re-search-forward md-roam-regex-tags-zettlr-style endpoint t)
+                 (let ((tag (match-string-no-properties 2)))
+                   (when tag
+                     (setq tags-list
+                           (append tags-list (list tag))))))
+               tags-list))))))
+
+(defun md-roam--extract-headlines (original-extract-headlines &optional file-path)
+  "Extract headline org-id's for backlinks if FILE-PATH is org file.
+This extraction is done via ORIGINAL-EXTRACT-HEADLINES fn:
+`org-roam--extract-headlines'. Return nil if not org files.
+It is meant to be used with `advice-add' :around."
+    
+  (let* ((file-path (or file-path
+                        (file-truename (buffer-file-name)))))
+    (if (md-roam--org-file-p file-path)
+        (apply original-extract-headlines file-path nil)
+      nil)))
+
+(advice-add 'org-roam--extract-headlines :around #'md-roam--extract-headlines)
+
 ;;;; Add advice to org-roam-db-build-cache
 
 (defun md-roam-add-message-to-db-build-cache (&optional force)
@@ -310,7 +446,6 @@ This is to simply indicate that md-roam is active. FORCE does not do anythying."
     (message "md-roam is active")))
 
 (advice-add 'org-roam-db-build-cache :before #'md-roam-add-message-to-db-build-cache)
-
 
 ;;;; Add advice to org-roam--get-roam-buffers
 ;;;; This is for org-roam-swtich-to-buffer